requirements.txt
Django>=2.1.2
mysqlclient==2.0.1
whitenoise==5.2.0
djangorestframework>=3.9.2
django-cors-headers==3.5.0
djangorestframework-simplejwt==4.4.0
# pytz==2018.8
#Successfully installed certifi-2020.6.20 chardet-3.0.4 django-oauth-toolkit-1.3.2 idna-2.10 oauthlib-3.1.0 requests-2.24.0 urllib3-1.25.10
#Successfully installed django-cors-middleware-1.5.0
#Successfully installed certifi-2020.6.20 chardet-3.0.4 django-oauth-toolkit-1.3.2 idna-2.10 oauthlib-3.1.0 requests-2.24.0 urllib3-1.25.10
#Successfully installed djangorestframework-simplejwt-4.4.0 pyjwt-1.7.1